Model returns lower case keys; incompatible w prev
It was a bit annoying to use CamelCase when everything else was lower
case, not to mention it created inconsistencies.

We thought about converting to underscore separation, but that would
likely have been significantly slower.

The model will now return a result that has all keys presented in lower
case.