Model properties are cleaned to match their schema
Some models vary slightly from query to query.  We want to try to have
some consistency, and it's nice to be able to leave out meaningless
attributes.

A model's result set will be limited only to the attributes defined in
the schema, and missing values will be filled in with a null.